restructure the full package
