+++ src/main/java/com/visualization/Model.java
@@ -1,4 +1,4 @@
-package vfp.main;
+package com.visualization;
